requirements.txt
requests
aiohttp
discord.py
colorlog
ephem
aiosqlite
